refactored code and extracted the animated card in a component so we could reuse it if needed
